+++ src/Haskellorls.hs
@@ -12,6 +12,7 @@ import qualified Data.Text.IO as T
 import qualified Haskellorls.Decorator as Decorator
 import qualified Haskellorls.Option as Option
 import qualified Haskellorls.Recursive as Recursive
+import qualified Haskellorls.Size.Utils as Size
 import qualified Haskellorls.Utils as Utils
 import qualified Options.Applicative as OA
 import qualified System.Exit as Exit
@@ -30,10 +31,11 @@ run args = do
   options <- argParser args
 
   isConnectedToTerminal <- IO.hIsTerminalDevice IO.stdout
+  blockSize <- Size.lookupBlockSize options
 
   if Option.version options
     then showVersion
-    else run' options {Option.toStdout = isConnectedToTerminal}
+    else run' options {Option.blockSize = blockSize, Option.toStdout = isConnectedToTerminal}
 
 run' :: Option.Option -> IO ()
 run' opt = do