+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - `-Z / --context` option.
 - `-N / --literal` option.
 - `-q / --hide-control-chars` and `--show-control-chars` options.
+- `-k / --kibibytes` option
+- Lookup `LS_BLOCK_SIZE` and `BLOCK_SIZE` environment variables to derive file block size.
 
 ### Changed